Convert to Dist::Zilla, including relocating tests.
This eliminates the Module::Build-based build system, and replaces it
with a Dist::Zilla based system. The changes here focus on a fairly
bare-minimum conversion.

One major change here is the reorganization of the tests, moving several
into a separate xt directory, and eliminating several completely, as
they are now auto-generated by dzil.